split major router parts into separate files
